app/utils/config.py
import os
from dotenv import load_dotenv

load_dotenv()

DATABASE_URL = os.getenv("DATABASE_URL", "sqlite:///./test.db")
REDIS_URL = os.getenv("REDIS_URL", "redis://localhost:6379")
SECRET_KEY = os.getenv("SECRET_KEY", "random-secret-key")
OPENAI_API_KEY = os.getenv("OPENAI_API_KEY", "your-default-openai-key")
