requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2
python-dotenv
passlib
bcrypt
pyjwt
openai
celery
redis
